requirements.txt
tensorflow-gpu
matplotlib
torch~=1.7.0+cu110
torchvision~=0.8.1+cu110
numpy~=1.18.5
Pillow~=8.0.1
opencv-python~=4.4.0.46